feat: changes in database for addtocompleted in seasonal
ID removed from PTW-CurrentSeason table, title is now primary key to serve for foreign key relation in SeasonalDetails.
Previously, deletes in the PTW-CurrentSeason table would shift the bottom entries up, causing a mismatch in the relationship between it and SeasonalDetails.
ID is now just order, to serve as the index for the order in which the entries appear in the sheet.